Windows build working, HHChannel fixed, pymoose cleaned up
This commit puts together several bug fixes and cleanups.

Meson-based build functional now.
- Windows MSVC Build works.
- Tested on WSL-Ubuntu as well
- Removed cmake build files (CMakeLists.txt)

HHChannel class hierarchy
- Original version of HHChannel was using the diamond of death
multiple inheritance. While linux/gcc was forgiving of this, on
Windows MSVC this was producing strange errors issue#483.
- Turns out this design was not essential and single inheritance along
the line ChanBase->ChanCommon->HHChannelBase->HHChannel sufficed.

pybind11 based pymoose

This code was quite dirty with static member functions under __Finfo__
which had no relations with the class. Reorganized the code here.

+++ basecode/Conv.h
@@ -570,7 +570,7 @@ template< class T > class Conv< vector< vector< T > > >
             ret.resize( numEntries );
             (*buf)++;
             for ( unsigned int i = 0; i < numEntries; ++i ) {
-                unsigned int rowSize = **buf;
+	      unsigned int rowSize = (unsigned int)**buf;
                 (*buf)++;
                 for ( unsigned int j = 0; j < rowSize; ++j )
                     ret[i].push_back( Conv< T >::buf2val( buf ) );
@@ -581,9 +581,9 @@ template< class T > class Conv< vector< vector< T > > >
         static void val2buf( const vector< vector< T > >& val, double**buf )
         {
             double* temp = *buf;
-            *temp++ = val.size();
+            *temp++ = (double)val.size();
             for( unsigned int i = 0; i < val.size(); ++i ) {
-                *temp++ = val[i].size();
+	      *temp++ = (double)val[i].size();
                 for ( unsigned int j = 0; j < val[i].size(); ++j ) {
                     Conv< T >::val2buf( val[i][j], &temp );
                 }
